requirements.txt
aioredis==1.3.1
amqp==2.6.1
asgiref==3.5.2
astroid==2.5
async-timeout==4.0.2
atomicwrites==1.4.0
attrs==22.1.0
autobahn==20.12.2
Automat==20.2.0
autopep8==1.5.4
Babel==2.10.3
backports.entry-points-selectable==1.1.1
beautifulsoup4==4.9.3
billiard==3.6.4.0
CacheControl==0.12.6
cachetools==4.2.1
cairocffi==1.3.0
CairoSVG==2.5.2
camelcase==0.2
celery==4.4.2
certifi==2020.6.20
cffi==1.14.6
channels==2.4.0
channels-redis==3.0.0
chardet==3.0.4
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
configparser==5.2.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
coverage==5.2
crayons==0.4.0
cryptography==3.4.7
cssselect2==0.4.1
cycler==0.11.0
daphne==2.5.0
defusedxml==0.7.1
Deprecated==1.2.13
distlib==0.3.2
distro==1.6.0
Django==3.0.8
django-appconf==1.0.4
django-axes==5.5.0
django-cacheops==5.0.1
django-celery-beat==2.0.0
django-celery-results==1.2.1
django-celerybeat-status==0.0.10
django-channels-presence==1.0.0
django-compressor==2.4
django-cors-headers==3.4.0
django-extensions==2.2.9
django-filter==2.3.0
django-grappelli==2.14.4
django-ipware==3.0.1
django-log-request-id==1.5.0
django-multiselectfield==0.1.12
django-redis==5.2.0
django-sass-processor==0.8
django-silk==4.1.0
django-simple-history==2.11.0
django-sortedm2m==3.0.2
django-timezone-field==4.0
django-utils-six==2.0
django-weasyprint==1.1.0
djangorestframework==3.12.4
djangorestframework-api-key==2.2.0
djangorestframework-oauth==1.1.0
djongo==1.3.6
drf-rw-serializers==1.0.4
drf-spectacular==0.22.1
drf-yasg==1.20.0
easygui==0.98.2
ecdsa==0.14.1
et-xmlfile==1.1.0
filelock==3.7.1
firebase-admin==5.0.2
flake8==3.8.4
flake8-polyfill==1.0.2
funcy==1.17
google-api-core==1.26.1
google-api-python-client==2.0.2
google-auth==1.28.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.3
google-cloud-core==1.6.0
google-cloud-firestore==2.3.1
google-cloud-storage==1.37.1
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
googledrivedownloader==0.4
gprof2dot==2022.7.29
greenlet==1.1.2
grpcio==1.47.0
hiredis==1.0.1
html5lib==1.1
httplib2==0.19.1
hyperlink==19.0.0
idna==2.10
importlib==1.0.4
incremental==21.3.0
inflection==0.5.1
isort==4.3.21
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==1.0.1
josepy==1.3.0
joypy==0.2.5
jsonfield==3.1.0
jsonschema==4.6.0
kiwisolver==1.4.4
kombu==4.6.11
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
matplotlib==3.4.2
mccabe==0.6.1
more-itertools==8.14.0
MouseInfo==0.1.3
msgpack==1.0.4
mysql-connector==2.2.9
numpy==1.20.2
oauthlib==3.1.0
opencv-contrib-python==4.5.3.56
opencv-python==4.5.3.56
openpyxl==3.0.7
packaging==20.4
pandas==1.2.4
pep8-naming==0.10.0
Pillow==8.2.0
platformdirs==2.2.0
playsound==1.2.2
pluggy==0.13.1
prompt-toolkit==3.0.29
proto-plus==1.20.6
protobuf==4.21.4
psycopg2-binary==2.8.6
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.52
pycodestyle==2.6.0
pycparser==2.20
pydocstyle==5.0.2
pydub==0.25.1
pyflakes==2.2.0
PyGetWindow==0.0.9
Pygments==2.6.1
PyHamcrest==2.0.2
PyJWT==2.4.0
pylint==2.5.2
pylint-celery==0.3
pylint-django==2.0.15
pylint-flask==0.6
pylint-plugin-utils==0.6
pymongo==4.2.0
PyMsgBox==1.0.9
pyOpenSSL==19.1.0
pyotp==2.4.0
pyparsing==2.4.7
pyperclip==1.8.2
Pyphen==0.10.0
PyRect==0.1.4
pyrsistent==0.18.1
PyScreeze==0.1.27
pytest==5.4.3
pytest-django==3.9.0
python-crontab==2.5.1
python-dateutil==2.8.1
python-jose==3.2.0
python-keycloak==0.22.0
PyTweening==1.0.3
pytz==2020.1
pywhatkit==4.8
PyYAML==5.3.1
quickchart.io==0.2.0
rcssmin==1.0.6
redis==3.2.0
requests==2.24.0
requests-oauthlib==1.3.0
requirements-detector==0.7
rjsmin==1.1.0
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scikit-learn==0.24.2
scipy==1.7.0
seaborn==0.11.1
selenium==3.141.0
sendotp==0.0.4
service-identity==21.1.0
six==1.16.0
sklearn==0.0
snowballstemmer==2.2.0
soupsieve==2.3.2.post1
splinter==0.16.0
SQLAlchemy==1.4.37
sqlparse==0.2.4
tabula==1.0.5
tabula-py==2.2.0
textmagic==2.0.3
threadpoolctl==2.2.0
tinycss2==1.1.0
tkcalendar==1.6.1
toml==0.10.2
Twisted==22.4.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typed-ast==1.4.1
typing_extensions==4.2.0
uritemplate==3.0.1
urllib3==1.25.9
uuid==1.30
vine==1.3.0
virtualenv==20.7.2
virtualenvwrapper-win==1.2.6
wcwidth==0.2.5
WeasyPrint==52.5
webdriver-manager==3.5.2
webencodings==0.5.1
wikipedia==1.4.0
wrapt==1.12.1
xlrd==1.2.0
XlsxWriter==1.4.3
zerosms==0.0.1
zipp==3.1.0
zope.interface==5.1.0
